Shift camelToSnake mapper to query builder
